[DEMO] Simplify imports and asset loading in sources
Simplify how TS code is imported in static JS file and how assets are loaded in page
The previous syntax work for the Vite dev server and demo bundling, but was not understood by IDE. We now use relative
paths that make code navigation possible.

Reorganize code to clarify what is internal and what is shared by example and test code
  - rename TS entry point for demo and test pages
  - move utils code in dedicated files
