+++ dev/public/diagram-navigation.html
@@ -6,8 +6,8 @@
 <head>
   <meta charset="UTF-8">
   <title>BPMN Visualization - Diagram Navigation</title>
-  <link rel="icon" href="static/img/favicon.png">
-  <link rel="stylesheet" href="static/css/test-page.css">
+  <link rel="icon" href="./static/img/favicon.png">
+  <link rel="stylesheet" href="./static/css/test-page.css">
   <style>
     .fit {
       background-color: #f9dfe8;
@@ -18,7 +18,7 @@
       width: 35px;
     }
   </style>
-  <script src="/dev/public/static/js/diagram-navigation.js" type="module"></script>
+  <script src="./static/js/diagram-navigation.js" type="module"></script>
 </head>
 <body>
 <div class="flex-container">
