+++ dev/public/elements-identification.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>BPMN Visualization - Elements Identification</title>
-    <link rel="icon" href="static/img/favicon.png">
+    <link rel="icon" href="./static/img/favicon.png">
     <style>
       #main-container {
         top: 140px;
@@ -120,7 +120,7 @@
         color: var(--color-flow) !important;
       }
     </style>
-    <script src="/dev/public/static/js/elements-identification.js" type="module"></script>
+    <script src="./static/js/elements-identification.js" type="module"></script>
 </head>
 <body>
     <div id="controls">
