+++ dev/public/index.html
@@ -5,9 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <title>BPMN Visualization Demo</title>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v6.1.1/css/all.css">
-    <link href="/dev/public/static/css/styles.css" rel="stylesheet">
-    <link rel="icon" href="static/img/favicon.png">
-    <script src="/dev/public/static/js/index.js" type="module"></script>
+    <link href="./static/css/styles.css" rel="stylesheet">
+    <link rel="icon" href="./static/img/favicon.png">
+    <script src="./static/js/index.js" type="module"></script>
 </head>
 <body class="bg-gray-800 font-sans leading-normal tracking-normal flex flex-col h-screen">
     <!--Nav-->
