+++ dev/public/lib-integration.html
@@ -3,8 +3,8 @@
 <head>
   <meta charset="UTF-8">
   <title>BPMN Visualization Lib Integration</title>
-  <link rel="icon" href="static/img/favicon.png">
-  <script src="/dev/public/static/js/lib-integration.js" type="module"></script>
+  <link rel="icon" href="./static/img/favicon.png">
+  <script src="./static/js/lib-integration.js" type="module"></script>
 </head>
 <body>
   <div id="bpmn-container-custom"></div>
