+++ dev/public/non-regression.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>BPMN Visualization Non Regression</title>
-    <link rel="icon" href="static/img/favicon.png">
+    <link rel="icon" href="./static/img/favicon.png">
     <style>
       #bpmn-container {
         top: 10px;
@@ -18,7 +18,7 @@
         font-weight: bold;
       }
     </style>
-    <script src="/dev/public/static/js/non-regression.js" type="module"></script>
+    <script src="./static/js/non-regression.js" type="module"></script>
 </head>
 <body>
     <div id="fetch-status"></div>
