+++ dev/public/overlays.html
@@ -6,14 +6,14 @@
 <head>
   <meta charset="UTF-8">
   <title>BPMN Visualization - Overlays</title>
-  <link rel="icon" href="static/img/favicon.png">
-  <link rel="stylesheet" href="static/css/test-page.css">
+  <link rel="icon" href="./static/img/favicon.png">
+  <link rel="stylesheet" href="./static/css/test-page.css">
   <style>
     .overlay {
       background-color: MintCream;
     }
   </style>
-  <script src="/dev/public/static/js/overlays.js" type="module"></script>
+  <script src="./static/js/overlays.js" type="module"></script>
 </head>
 <body>
 <div class="flex-container">
