+++ dev/public/static/js/lib-integration.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { BpmnVisualization } from '/src/bpmn-visualization.ts';
+import { BpmnVisualization } from '../../../../src/bpmn-visualization';
 
 const bpmnVisualizationIntegration = new BpmnVisualization({ container: 'bpmn-container-custom' });
 bpmnVisualizationIntegration.load(bpmnDefaultContent());
